requirements.txt
streamlit
cloudinary
sqlalchemy
psycopg2-binary
pandas
pdfplumber
python-docx
python-dotenv